examples/fastapi-tanstack-api/pyproject.toml
[project]
name = "fastapi-tanstack-example"
version = "0.0.1"
description = "FastAPI + TanStack example"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.5.9",
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "greenlet>=3.3.2",
    "openai-agents[litellm,sqlalchemy]>=0.10.5",
    "python-dotenv>=1.2.2",
    "scalar-fastapi>=1.8.0",
    "tavily-python>=0.7.22",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
exclude = []
